+++ src/model/product/item.h
@@ -9,17 +9,17 @@ using Items = std::vector<Item>;
 struct Item{
     int id;
     int count;
-    long long price;
+    int price;
 
     Item(): id(-1),count(-1),price(-1) {}
-    Item(int id, int count, long long price)
+    Item(int id, int count, int price)
         : id(id), count(count), price(price) {}
 
     int getId() {return id;}
     int getCount() {return count;}
-    long long getPrice() {return price;}
+    int getPrice() {return price;}
 
     void setId(int id) {this->id = id;}
     void setCount(int count) {this->count = count;}
-    void setPrice(long long price) {this->price = price;}
+    void setPrice(int price) {this->price = price;}
 };
