+++ src/model/user/userInfo.h
@@ -9,7 +9,7 @@ using std::string;
 class UserInfo {
     UserRole role;
     string name;
-    long long point;
+    int point;
 
     std::string infoText;
 
@@ -18,12 +18,12 @@ public:
 
     void setRole(UserRole role) {this->role = role;}
     void setName(string name) {this->name = name;}
-    void setPoint(long long point) {this->point = point;}
+    void setPoint(int point) {this->point = point;}
     void setInfoText(string infoTet) {this->infoText = infoTet;}
 
     UserRole getRole() {return role;}
     string getName() {return name;}
-    long long getPoint() {return point;}
+    int getPoint() {return point;}
     string getInfoText() {return infoText;}
 
     UserInfo* operator=(User& user);
